blscfg: Always look for BLS snippets in the root device
The blscfg command trying to load the BLS snippets from different devices
is a left over from when these where located in the ESP in the EFI case.

But now that are always located in the /boot directory, the BLS snippets
have to be loaded from the root device just like the other files that are
in the /boot directory (kernel and initrd images).

This only worked for legacy BIOS because the boot and root variables were
set to the same value. But doesn't work when trying to use a common GRUB
config file that could be shared between an EFI and legacy BIOS install.

That kind of configuration is needed for example on hybrid images that are
booted using either EFI or legacy BIOS firmwares.

Resolves: rhbz#1874470

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>