blscfg: Don't hardcode an env var as fallback for the BLS options field
If the BLS fragments don't have an options field or if this was set to an
environment variable that was not defined in the grubenv file, the blscfg
module searches for a default_kernelopts variable that is defined in the
grub.cfg file.

But the blscfg module shouldn't hardcode fallbacks variables and instead
this logic should be handled in the GRUB config file itself.

Also, add a comment explaining where the kernelopts variable is supposed
to be defined and what is the process for the user to change its value.

Resolves: rhbz#1874469

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>